requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.48.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.4.1
dill==0.4.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.7
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
peft==0.18.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
trl==0.25.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
